Сделал часть ДЗ
